requirements.txt
# OpenRouter API (Gemini 2.5 Pro)
# Utilise l'API compatible OpenAI via requests

# Framework API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Traitement des données
pandas>=2.1.0
numpy>=1.24.0

# Text-to-Speech
gtts>=2.4.0
pydub>=0.25.1

# Autres utilitaires
python-dotenv>=1.0.0
requests>=2.31.0